% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_scale_wells}
\alias{calculate_scale_wells}
\title{Title}
\usage{
calculate_scale_wells(plate_img, well_properties, plate_type)
}
\arguments{
\item{well_properties}{}
}
\value{

}
\description{
Title
}
