% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_setup_data}
\alias{combine_setup_data}
\title{Function to combine data with setup info}
\usage{
combine_setup_data(setup, data)
}
\arguments{
\item{setup}{A file with identification of what is in each well}

\item{data}{A file with data from each well}
}
\value{
A dataframe with combined data from setup and input
}
\description{
Function to combine data with setup info
}
