% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_img_values}
\alias{extract_img_values}
\title{Title}
\usage{
extract_img_values(img_file, p, invert, well_frame, pp, scale_properties)
}
\arguments{
\item{scale_properties}{}
}
\value{

}
\description{
Title
}
