% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_well_values}
\alias{extract_well_values}
\title{Title}
\usage{
extract_well_values(img, well_position, px_per_mm, inter_well_space)
}
\arguments{
\item{inter_well_space}{}
}
\value{

}
\description{
Title
}
