% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_img_settings}
\alias{get_img_settings}
\title{Title}
\usage{
get_img_settings(img_idx, img_settings_df)
}
\arguments{
\item{img_settings_list}{e.g. list(c(panel="blue", exposure=20000, intensity=1))}
}
\value{

}
\description{
Title
}
