% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_scale_and_location}
\alias{get_scale_and_location}
\title{Title}
\usage{
get_scale_and_location(plate_img, num_wells, experiment_type, plate_type)
}
\arguments{
\item{plate_type}{}
}
\value{

}
\description{
Title
}
