% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateUtilities.R
\name{getWellIds}
\alias{getWellIds}
\title{Returns a character vector of well IDs (e.g. A01..B10..H12) of length 12,
             24, 48, 96, or 384 wells.}
\usage{
getWellIds(plateSize)
}
\arguments{
\item{plateSize}{12, 24, 48, 96, or 384 wells}
}
\value{
A character vector of well IDs (e.g. A01..B05..H12) of length 12,
             24, 48, 96, or 384
}
\description{
Returns a character vector of well IDs (e.g. A01..B10..H12) of length 12,
             24, 48, 96, or 384 wells.
}
\examples{
getWellIds(96)
}

