% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{interactive_scale_wells}
\alias{interactive_scale_wells}
\title{Title}
\usage{
interactive_scale_wells(plate_img, pp, scale_properties)
}
\arguments{
\item{scale_properties}{}
}
\value{

}
\description{
Title
}
