% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process_img_dir}
\alias{process_img_dir}
\title{Process a folder of images}
\usage{
process_img_dir(
  dir_path,
  align_filename,
  invert = F,
  layout_csv,
  experiment_type = "colony",
  plate_type = "6-well",
  num_wells = 384,
  px_per_mm = NA,
  tl_corner = NA,
  in_parallel = FALSE,
  img_settings_list
)
}
\arguments{
\item{dir_path}{path to directory containing plate images (.png images only)}

\item{align_filename}{filename of image used to align well grid}

\item{invert}{Boolean flag. Set to TRUE if the images are upside-down}

\item{layout_csv}{.csv with metadata for each position on the plate}

\item{experiment_type}{"colony" or "lawn"}

\item{plate_type}{"6-well" or "1-well"}

\item{num_wells}{Plate size used for output positions: 96, 384 or 1536}

\item{img_settings_list}{}
}
\value{

}
\description{
Process a folder of images
}
