% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_plate}
\alias{read_plate}
\title{Function convert plate format into machine readable table}
\usage{
read_plate(df, cols = "plate", n = 1)
}
\arguments{
\item{df}{Dataframe in plate format (see below for csv-like example)

plate, 1, 2, 3, 4, ...
A, ...
B, ...
C, ...
..., ...}

\item{cols}{String with name of plate in top-left corner of plate (like
"plate" in csv-like example above)}

\item{n}{Integer to identify plate (i.e. 1, 2, 3...)}
}
\value{
Dataframe with microtiter plate content
}
\description{
Function convert plate format into machine readable table
}
