% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_plate_xlsx}
\alias{read_plate_xlsx}
\title{Function convert plate format into machine readable table}
\usage{
read_plate_xlsx(fp, sheet, cols = "plate", n = 1)
}
\arguments{
\item{fp}{File path to xlsx file with plate data}

\item{sheet}{Worksheet number (1, 2, 3, 4, ...) in excel worksbook}

\item{cols}{String with name of plate in top-left corner of plate (like
"plate" in csv-like example above)}

\item{n}{Integer to identify plate (i.e. 1, 2, 3...)}
}
\value{
Dataframe with microtiter plate content
}
\description{
Function convert plate format into machine readable table
}
