% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_plate.R
\name{add_plate}
\alias{add_plate}
\title{Reads plate layouts and adds them to an existing data frame.}
\usage{
add_plate(file, data, well_ids_column, plate_size = 96)
}
\arguments{
\item{file}{A character vector with the path of a .csv file formatted as
as described in \code{\link{read_plate}}.}

\item{data}{The data frame to merge the new annotations into. Must contain
a column with well names.}

\item{well_ids_column}{The name of the column in \code{data} containing the
well IDs.}

\item{plate_size}{The number of wells in the plate. Must be 12, 24, 48, 96 or
384. Default 96.}
}
\value{
Returns data with as many new columns as plates in \code{file}.
Empty wells are indicated with NA.
}
\description{
Converts data from a microtiter plate layout to a data frame with one well
per row and merges it into an existing data frame by well name.
}
\details{
If data contains more wells than in \code{file}, NA will be added to the
merged column for those wells. If the file contains more wells than
\code{data}, an error will result.
}

