% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_scale_and_location}
\alias{get_scale_and_location}
\title{Title}
\usage{
get_scale_and_location(
  plate_img,
  num_wells,
  experiment_type,
  plate_type,
  deg_rotation,
  well_ratio
)
}
\arguments{
\item{deg_rotation}{}
}
\value{

}
\description{
Title
}
