% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process_img_dir}
\alias{process_img_dir}
\title{Process a folder of images}
\usage{
process_img_dir(
  dir_path,
  align_filename,
  invert = F,
  rotate = F,
  layout_csv,
  experiment_type = "colony",
  normalise = F,
  plate_type = "6-well",
  num_wells = 384,
  px_per_mm = NA,
  tl_corner = NA,
  well_ratio = NA,
  in_parallel = FALSE,
  img_settings_list
)
}
\arguments{
\item{dir_path}{path to directory containing plate images (.png images only)}

\item{align_filename}{filename of image used to align well grid}

\item{invert}{Boolean flag. Set to {TRUE} if the plate is inverted in the imager}

\item{rotate}{Boolean flag. Set to {TRUE} if the images are upside-down i.e. row A on bottom of image}

\item{layout_csv}{.csv with metadata for each position on the plate}

\item{experiment_type}{"colony" or "lawn"}

\item{normalise}{Attempts to remove background. Negates pixel values in the first iteration of images, from all images.}

\item{plate_type}{"6-well" or "1-well"}

\item{num_wells}{Plate size used for output positions: 96, 384 or 1536}

\item{px_per_mm}{If you already know the `px_per_mm` value (from previous alignment)}

\item{tl_corner}{If you already know the `tl_corner` values (from previous alignment)}

\item{well_ratio}{Proportion of space taken up by each "well". Can be scaled interactively.}

\item{in_parallel}{process images in parallel (only available for R >= 4.0)}

\item{img_settings_list}{a list of named vectors with meta-data for images in the order they were taken i.e.
`img_settings_list = list(c(panel="blue", exposure=20000, intensity=1),c(panel="red", exposure=4000, intensity=0.7))`}
}
\value{

}
\description{
Process a folder of images
}
