% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summarise_image}
\alias{summarise_image}
\title{Title}
\usage{
summarise_image(
  img_file,
  normalise,
  invert,
  rotate,
  well_frame,
  img_idx,
  img_settings_df,
  circ_stencil
)
}
\arguments{
\item{rotate}{}
}
\value{

}
\description{
Title
}
