% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view_plate.R
\name{view_plate}
\alias{view_plate}
\title{Displays the data in the form of a microtiter plate.}
\usage{
view_plate(data, well_ids_column, column_to_display, plate_size = 96)
}
\arguments{
\item{data}{A data frame containing the data}

\item{well_ids_column}{The name of the column in \code{data} containing the
well IDs.}

\item{column_to_display}{The column containing the data to display.}

\item{plate_size}{The number of wells in the plate. Must be 12, 24, 48, 96 or
384. Default 96.}
}
\value{
A data frame representing the data in \code{column_to_display} as
though laid out on a microtiter plate with \code{plate_size} wells.
}
\description{
Displays the data in the form of a microtiter plate.
}

