% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_plate.R
\name{view_plate}
\alias{view_plate}
\title{Displays the data in the form of a microtiter plate.}
\usage{
view_plate(data, well_ids_column, column_to_display, plate_size = 96)
}
\arguments{
\item{data}{A data frame containing the data}

\item{well_ids_column}{The name of the column in \code{data} containing the 
well IDs.}

\item{column_to_display}{The column containing the data to display.}

\item{plate_size}{The number of wells in the plate. Must be 12, 24, 48, 96 or
384. Default 96.}
}
\value{
A depiction of the data in \code{column_to_display} as 
though laid out on a microtiter plate with \code{plate_size} wells.
}
\description{
Displays the data in the form of a microtiter plate.
}
\examples{
# Get some tidy data
file_path <- system.file("extdata", "example-1.csv", package = "plateR")
data <- read_plate(file = file_path, well_ids_column = "Wells")
head(data)

# See which wells got which drug
view_plate(data, "Wells", "Drug", 96)

# See concentration of the drugs across the plate
view_plate(data, "Wells", "Concentration", 96)
}

